refactor(cumulative): added a custom hook for cumulative code
