+++ src/components/cell-list.css
@@ -1,0 +1,7 @@
+.cell-list {
+  margin: 0 25px 50vh;
+}
+
+.react-draggable-transparent-selection .cell-list {
+  margin-bottom: 100vh;
+}
