+++ src/components/cell-list.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useTypedSelector } from '../hooks';
 import CellListItem from './cell-list-item';
 import AddCell from './add-cell';
+import './cell-list.css';
 
 const CellList: React.FC = () => {
   const cells = useTypedSelector(({ cells: { order, data } }) =>
@@ -15,7 +16,7 @@ const CellList: React.FC = () => {
     </React.Fragment>
   ));
   return (
-    <div>
+    <div className="cell-list">
       <AddCell forceVisible={cells.length === 0} prevCellId={null} />
       {renderedCells}
     </div>
