+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 import { useActions } from './use-actions';
 import { useTypedSelector } from './used-typed-selector';
+import { useCumulativeCode } from './use-cumulative-code';
 
-export { useActions, useTypedSelector };
+export { useActions, useTypedSelector, useCumulativeCode };
